refactor: move helper in separate file
