+++ src/shell/mod.rs
@@ -2,5 +2,6 @@ mod error;
 mod state;
 
 pub mod command;
+pub mod helper;
 
 pub use state::State;
